requirement.txt
Flask
flask-cors
python-docx
pdfplumber
